chore: refactor kube job columns
Refactor the Kubernetes Job columns with /columns and props.ts.

Part of #11035.
