+++ packages/renderer/src/lib/job/JobList.svelte
@@ -8,10 +8,10 @@ import JobIcon from '../images/JobIcon.svelte';
 import NameColumn from '../kube/column/Name.svelte';
 import StatusColumn from '../kube/column/Status.svelte';
 import KubernetesObjectsList from '../objects/KubernetesObjectsList.svelte';
+import ActionsColumn from './columns/Actions.svelte';
+import CompletionsColumn from './columns/Completions.svelte';
+import ConditionsColumn from './columns/Conditions.svelte';
 import { JobUtils } from './job-utils';
-import JobColumnActions from './JobColumnActions.svelte';
-import JobColumnCompletions from './JobColumnCompletions.svelte';
-import JobColumnConditions from './JobColumnConditions.svelte';
 import JobEmptyScreen from './JobEmptyScreen.svelte';
 import type { JobUI } from './JobUI';
 
@@ -45,12 +45,12 @@ let ageColumn = new TableColumn<JobUI, Date | undefined>('Age', {
 });
 
 let conditionColumn = new TableColumn<JobUI>('Conditions', {
-  renderer: JobColumnConditions,
+  renderer: ConditionsColumn,
   comparator: (a, b): number => a.condition.localeCompare(b.condition),
 });
 
 let completionColumn = new TableColumn<JobUI>('Completions', {
-  renderer: JobColumnCompletions,
+  renderer: CompletionsColumn,
   comparator: (a, b): number => a.succeeded.toString().localeCompare(b.succeeded.toString()),
 });
 
@@ -60,7 +60,7 @@ const columns = [
   conditionColumn,
   completionColumn,
   ageColumn,
-  new TableColumn<JobUI>('Actions', { align: 'right', renderer: JobColumnActions }),
+  new TableColumn<JobUI>('Actions', { align: 'right', renderer: ActionsColumn }),
 ];
 
 const row = new TableRow<JobUI>({ selectable: (_job): boolean => true });
