+++ packages/renderer/src/lib/job/columns/Actions.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+import JobActions from '../JobActions.svelte';
+import type { Props } from './props';
+
+let { object }: Props = $props();
+</script>
+
+<JobActions job={object} />
