+++ packages/renderer/src/lib/job/columns/Completions.spec.ts
@@ -21,10 +21,10 @@ import '@testing-library/jest-dom/vitest';
 import { render, screen } from '@testing-library/svelte';
 import { expect, test } from 'vitest';
 
-import JobColumnActions from './JobColumnActions.svelte';
-import type { JobUI } from './JobUI';
+import type { JobUI } from '../JobUI';
+import Completions from './Completions.svelte';
 
-test('Expect job buttons', async () => {
+test('Expect simple column styling', async () => {
   const job: JobUI = {
     uid: '123',
     name: 'my-job',
@@ -36,8 +36,9 @@ test('Expect job buttons', async () => {
     condition: 'running',
   };
 
-  render(JobColumnActions, { object: job });
+  render(Completions, { object: job });
 
-  const buttons = await screen.findAllByRole('button');
-  expect(buttons).toHaveLength(1);
+  const text = screen.getByText(job.succeeded + ' / ' + job.completions);
+  expect(text).toBeInTheDocument();
+  expect(text).toHaveClass('text-[var(--pd-table-body-text)]');
 });
